Remove required on the currency and fallback to company's currency...
... on the policy level if it is empty.

Also, it fixes a bug when a credit line had no currency and the level
had the company's currency. As False is different than any currency,
it entered in res_currency.compute() with a False currency. It must
just fallback on the company's currency if the credit line's currency
is empty.