[ADD] Auto process lower levels
Add the possibility to mark a policy so that processing a line generated by this policy
will also process lower level lines for the same partner (and policy).

Also add a filter for credit control line tree view in order to only display
highest level lines for each partner and policy.