[IMP] account_financial_risk: Add specific overpassing group
This commit adds an specific group for overpassing partner credit limit exceptions.
In some situations, not only account managers shoud be allowed to overpass exception, depending on the company policies. Merging this code will not affect current instances, as long as account managers will be added to the new group.