[FIX] account_financial_risk: Fix tests fixing a date
There's a part of the risk computation that depends on current date,
and this makes the test to fail when time has past.

We force the maturity date on the invoice we are testing for always
having the field properly computed. An alternative would be to use
``freezegun``, but that means more changes in the test suite.