[FIX] account_financial_risk: Don't store company_currency_id
The field is not populated on proper time (only once saved) when
recomputing dynamic lines if it's stored, provoking an error.

Moreover, making it stored seems a random decision, as it's not
benefitting the computation of the rest of the fields, as they are not
stored as well.

Being on it, the depends of the other computed field has been adjusted
to the proper field. Needing the other depend to work was a symptom of
this issue that was ignored.

TT33885