[FIX] sale_financial_risk: fix test
The test_manual_currency_risk_not_exceeded test that assesses that the
manual currency risk has not been passed fails when another module changes
the company's currency. The reason for the failure is that if the company
currency matches the currency of the sell order it does not convert
correctly and in that case it interprets the limit within the same currency.
For example, if the company currency is set as EUR and the sales order
is placed in EUR, the conversion of the 100€ is not done and therefore
exceeds the limit set to 99. To solve this, the company currency is
first checked to set a different currency in the order. On the other
hand, the tests should be frozen at a certain date to prevent the
exchange rate of the exchange currency from going up or down thus
avoiding errors when running the test, which should check that it has
not been exceeded.